+++ usr/src/lib/libdladm/common/libdlmgmt.h
@@ -98,7 +98,7 @@ typedef struct dlmgmt_door_setattr_s {
 	int			ld_cmd;
 	int			ld_confid;
 	char			ld_attr[MAXLINKATTRLEN];
-	size_t			ld_attrsz;
+	uint32_t		ld_attrsz;
 	dladm_datatype_t	ld_type;
 	char			ld_attrval[MAXLINKATTRVALLEN];
 } dlmgmt_door_setattr_t;
@@ -152,7 +152,7 @@ typedef struct dlmgmt_openconf_retval_s {
 
 typedef struct dlmgmt_getconfsnapshot_retval_s {
 	uint_t		lr_err;
-	size_t		lr_nvlsz;
+	uint32_t	lr_nvlsz;
 	/* buffer for nvl */
 } dlmgmt_getconfsnapshot_retval_t;
